78693180/requirements.txt
scrapy-playwright==0.0.34
beautifulsoup4==4.12.3
